CR feedback - move test into an existing test
